Move video classes to one package
