+++ collect_app/src/test/java/org/odk/collect/android/widgets/ExVideoWidgetTest.java
@@ -27,6 +27,7 @@ import org.odk.collect.android.widgets.support.FakeQuestionMediaManager;
 import org.odk.collect.android.widgets.support.FakeWaitingForDataRegistry;
 import org.odk.collect.android.widgets.utilities.FileRequester;
 import org.odk.collect.android.widgets.utilities.QuestionFontSizeUtils;
+import org.odk.collect.android.widgets.video.ExVideoWidget;
 import org.odk.collect.androidshared.system.IntentLauncher;
 import org.robolectric.shadows.ShadowToast;
